Se crean archivos en local y remoto

+++ CrearArchivos.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+from service_drive import obtener_servicio
 import os
 
 
@@ -8,29 +10,67 @@ def tipos_archivos() -> None:
           "3) Crear un archivo .csv\n")
 
 
+def crear_carpeta_drive(nombre_carpeta: str):
+    archivo_metadata = {
+        'name': nombre_carpeta,
+        'mimeType': 'application/vnd.google-apps.folder'
+    }
+    obtener_servicio().files().create(body= archivo_metadata, fields = 'id').execute()
+    print("Se creo su carpeta")
+
+
 def crear_carpeta() -> None:
     # Crea carpetas, falta que naveguen en el escritorio
     nombre_carpeta = input("Ingrese el nombre de la carpeta: ")
     if os.path.isdir(nombre_carpeta):
         print("la carpeta existe")
     else:
+        crear_carpeta_drive(nombre_carpeta)
         os.mkdir(nombre_carpeta)
+
+
+def crear_archivo_txt_drive(nombre_archivo):
+    archivo_metadata = {
+        'name': nombre_archivo,
+        'mimeType': 'text/plain'
+    }
+    obtener_servicio().files().create(body = archivo_metadata, fields = 'id').execute()
+    print("Se creo su archivo .txt")
 
 
 def crear_archivo_txt() -> None:
     # Crea archivos .txt falta la navegacion
     nombre_archivo = input("Ingrese el nombre del nuevo archivo: ")
     nombre_archivo += ".txt"
-    nuevo_archivo = open(nombre_archivo, "w")
-    nuevo_archivo.close()
+    archivo = Path(nombre_archivo)
+    if archivo.is_file():
+        print("Ese nombre de archivo ya existe")
+    else:
+        crear_archivo_txt_drive(nombre_archivo)
+        nuevo_archivo = open(nombre_archivo, "w")
+        nuevo_archivo.close()
+
+
+def crear_archivo_csv_drive(nombre_archivo):
+    archivo_metadata = {
+        'name': nombre_archivo,
+        'mimeType': 'text/plain'
+    }
+    obtener_servicio().files().create(body = archivo_metadata, fields = 'id').execute()
+    print("Se creo su archivo .csv")
 
 
 def crear_archivo_csv() -> None:
     # Crea archivos .csv, falta la navegacion
     nombre_archivo = input("Ingres el nombre del nuevo archivo: ")
     nombre_archivo += ".csv"
-    nuevo_archivo = open(nombre_archivo, "w")
-    nuevo_archivo.close()
+    archivo = Path(nombre_archivo)
+    if archivo.is_file():
+        print("Ese nombre de archivo ya existe")
+    else:
+        crear_archivo_csv_drive(nombre_archivo)
+        nuevo_archivo = open(nombre_archivo, "w")
+        nuevo_archivo.close()
 
 
 def creacion_archivos() -> None:
